initial non-durable reads commit flush option
we mark the transactional internally as committed
before we sync the WAL log. this creates a read
anomaly where another transaction may see results
from this transaction which are later lost if the
WAL fails to reach the disk. this is a possible
alternative to synchronous_commit = off since it
allows writers to wait for the transaction to 
become durable but it does not block other
transactions from seeing the results of the 
transaction. in particular if there is high
contention on a tuple then writes will be
serialized around fsync'ing the WAL log. this
optimization allows the WAL log writes to be
batched in this scenario.